PR #47 Additions

+++ MidasCivil_Adapter/CRUD/Delete/Elements/Element.cs
@@ -14,44 +14,46 @@ namespace BH.Adapter.MidasCivil
 
             if (ids!=null)
             {
-                List<string> stringIndex = ids.Cast<string>().ToList();
-
-                List<int> indices = ids.Cast<int>().ToList();
-
                 string path = directory + "\\" + "ELEMENT" + ".txt";
 
-                List<string> elements = File.ReadAllLines(path).ToList();
-                List<int> elementIndexes = new List<int>();
-
-                foreach (string element in elements)
+                if(File.Exists(path))
                 {
-                    if (element.Contains("*") || element.Contains(";"))
+                    List<string> stringIndex = ids.Cast<string>().ToList();
+
+                    List<int> indices = ids.Cast<int>().ToList();
+
+                    List<string> elements = File.ReadAllLines(path).ToList();
+                    List<int> elementIndexes = new List<int>();
+
+                    foreach (string element in elements)
                     {
-                        int clone = 0;
-                        elementIndexes.Add(clone);
+                        if (element.Contains("*") || element.Contains(";"))
+                        {
+                            int clone = 0;
+                            elementIndexes.Add(clone);
+                        }
+                        else
+                        {
+                            int clone = int.Parse(element.Split(',')[0].Replace(" ", ""));
+                            elementIndexes.Add(clone);
+                        }
                     }
-                    else
+
+                    foreach (int index in indices)
                     {
-                        int clone = int.Parse(element.Split(',')[0].Replace(" ", ""));
-                        elementIndexes.Add(clone);
+                        if (elementIndexes.Contains(index))
+                        {
+                            int elementIndex = elementIndexes.IndexOf(index);
+                            elements[elementIndex] = "";
+                        }
                     }
+
+                    elements = elements.Where(x => !string.IsNullOrWhiteSpace(x)).ToList();
+
+                    File.Delete(path);
+                    File.WriteAllLines(path, elements.ToArray());
                 }
-
-                foreach (int index in indices)
-                {
-                    if (elementIndexes.Contains(index))
-                    {
-                        int elementIndex = elementIndexes.IndexOf(index);
-                        elements[elementIndex] = "";
-                    }
-                }
-
-                elements = elements.Where(x => !string.IsNullOrWhiteSpace(x)).ToList();
-
-                File.Delete(path);
-                File.WriteAllLines(path, elements.ToArray());
             }
-
             return success;
         }
     }
